+++ isaaclab_arena/relations/no_overlap_mesh.py
@@ -17,7 +17,8 @@ from isaaclab_arena.relations.mesh_pair_cache import MeshPairCache, MeshPairEntr
 from isaaclab_arena.relations.relation_solver_state import RelationSolverState
 from isaaclab_arena.relations.warp_mesh_manager import WarpMeshAndSphereCache
 from isaaclab_arena.relations.warp_sdf_kernels import clamp_sdf_sentinel, multi_mesh_sdf
-from isaaclab_arena.utils.pose import Pose, rotate_points_by_yaw_batch, yaw_from_quat_xyzw
+from isaaclab_arena.utils.pose import Pose
+from isaaclab_arena.utils.yaw import rotate_points_by_yaw_batch, yaw_from_quat_xyzw
 
 if TYPE_CHECKING:
     from isaaclab_arena.assets.object_base import ObjectBase
